backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
python-multipart
sqlalchemy
psycopg[binary]
pgvector
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
pypdf
